# api.yaml
...
info:
  description: Task for LongShot AI Assingment Backend. Backend API to manage and categorize stock for a small-scale grocery store.
  title: LongShot AI Assingment Backend
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /storageSpaces/create:
    post:
      summary: Create a storage space
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the storage space
                maximumLimit:
                  type: integer
                  description: The maximum limit of the storage space
                refrigeration:
                  type: boolean
                  description: Whether the storage space requires refrigeration or not
              required:
                - name
                - maximumLimit
                - refrigeration
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Storage space created
                  data:
                    $ref: '#/components/schemas/StorageSpace'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
                  error:
                    type: object
                    description: The error object
                    properties:
                      message:
                        type: string
                        description: Error message
      tags:
        - Storage Spaces
...
servers:
  - description: Local server
    url: http://localhost:5000/